Tokens
Added a player token and event tokens.  I also fixed part of the doors so there is a door that always leads to the next tile.  I need to make sure that the new tile's doors line up with the previous token and to add neighbors so that all tile doors match.